OBDX-16 Create a method to send data on picked, packed, dispatched and on quantity change
